git-svn-id: http://svn.l2jfree.com/l2jfree@6498 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/templates/item/L2Item.java
@@ -140,7 +140,7 @@ public abstract class L2Item implements FuncOwner
 	private final int					_materialType;
 	private final int					_crystalType;																	// default to none-grade
 	private final int					_duration;
-	private final int					_time;
+	private final long					_time;
 	private final int					_bodyPart;
 	private final int					_referencePrice;
 	private final int					_crystalCount;
@@ -220,7 +220,7 @@ public abstract class L2Item implements FuncOwner
 	 * Returns the time of the item
 	 * @return int
 	 */
-	public final int getTime()
+	public final long getTime()
 	{
 		return _time;
 	}
